# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Theke API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/resources/{id}/restore": {
      "post": {
        "operationId": "restoreResource",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "201": { "description": "Recurso restaurado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResourceResponse" } } } }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
...
    "/v1/resources": {
      "get": {
        "operationId": "listResources",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "note",
                "file",
                "link"
              ]
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["active", "archived"],
              "default": "active"
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "Coincidencia parcial por título, alias, descripción o contenido de nota",
            "schema": {
              "type": "string",
              "maxLength": 160
            }
          },
          {
            "name": "projectId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "folderId",
            "in": "query",
            "description": "Requiere projectId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "libraryFolderId",
            "in": "query",
            "description": "UUID de carpeta o root para archivos sin carpeta",
            "schema": { "type": "string" }
          },
          {
            "name": "tag",
            "in": "query",
            "schema": { "type": "string", "maxLength": 80 }
          }
        ],
        "responses": {
          "200": {
            "description": "Página de recursos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/v1/resources/property-definitions": {
      "get": {
        "operationId": "listResourcePropertyDefinitions",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Tipos de propiedades definidos en la Cuenta", "content": { "application/json": { "schema": { "type": "object", "required": ["data"], "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/ResourcePropertyDefinition" } } } } } } } }
      }
    },
    "/v1/resources/{id}/metadata": {
      "patch": {
        "operationId": "updateResourceMetadata",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateResourceMetadataInput" } } } },
        "responses": { "200": { "description": "Metadatos guardados", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResourceMetadataResponse" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "404": { "$ref": "#/components/responses/NotFound" }, "409": { "description": "Versión concurrente o tipo incompatible", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } } } }
      }
    },
    "/v1/resources/{id}/references": {
      "get": {
        "operationId": "getResourceReferences",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Enlaces de la versión actual y referencias entrantes actuales", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResourceReferencesResponse" } } } }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
    "/v1/resources/{id}": {
      "get": {
        "operationId": "getResource",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle del recurso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/v1/resources/{id}/access": {
      "get": {
        "operationId": "getResourceAccess",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "mode",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "inline",
                "download"
              ],
              "default": "inline"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "URL privada temporal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceAccessResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/v1/resources/{id}/accessibility": {
      "patch": {
        "operationId": "updateResourceAccessibility",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessibilityInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Accesibilidad actualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
    "responses": {
      "BadRequest": {
        "description": "Solicitud inválida",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Sesión inválida",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "No encontrado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {},
          "requestId": {
            "type": "string"
          }
        }
      },
      "ApiErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        }
      },
...
      "Resource": {
        "type": "object",
        "required": [
          "id",
          "title",
          "type",
          "origin",
          "status",
          "archivedAt",
          "updatedAt",
          "libraryFolderId",
          "aliases",
          "tags",
          "properties",
          "versionId",
          "accessibilityRequired",
          "accessibilityMissing"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "aliases": { "type": "array", "items": { "type": "string" } },
          "tags": { "type": "array", "items": { "type": "string" } },
          "properties": { "type": "object", "additionalProperties": { "oneOf": [{ "type": "string" }, { "type": "number" }, { "type": "boolean" }, { "type": "array", "items": { "type": "string" } }] } },
          "versionId": { "type": ["string", "null"], "format": "uuid" },
          "type": {
            "type": "string",
            "enum": [
              "note",
              "file",
              "link"
            ]
          },
          "mediaType": {
            "type": [
              "string",
              "null"
            ]
          },
          "byteSize": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "origin": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ready",
              "archived"
            ]
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "libraryFolderId": { "type": ["string", "null"], "format": "uuid" },
          "archivedAt": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "content": {
            "type": "string"
          },
          "accessibilityText": {
            "type": [
              "string",
              "null"
            ]
          },
          "accessibilityRequired": {
            "type": "boolean"
          },
          "accessibilityMissing": {
            "type": "boolean"
          },
          "url": {
            "type": ["string", "null"],
            "format": "uri"
          },
          "previewImageUrl": {
            "type": ["string", "null"],
            "format": "uri"
          },
          "metadataStatus": {
            "type": ["string", "null"],
            "enum": ["pending", "ready", "failed", null]
          }
        }
      },
...
      "ResourcePropertyDefinition": { "type": "object", "required": ["key", "type"], "properties": { "key": { "type": "string" }, "type": { "type": "string", "enum": ["text", "list", "number", "checkbox", "date", "datetime"] } } },
      "ResourcePropertyInput": { "type": "object", "required": ["key", "type", "value"], "properties": { "key": { "type": "string", "maxLength": 40 }, "type": { "type": "string", "enum": ["text", "list", "number", "checkbox", "date", "datetime"] }, "value": { "oneOf": [{ "type": "string" }, { "type": "number" }, { "type": "boolean" }, { "type": "array", "items": { "type": "string" } }] } } },
      "UpdateResourceMetadataInput": { "type": "object", "required": ["aliases", "tags", "properties", "expectedUpdatedAt"], "properties": { "aliases": { "type": "array", "maxItems": 20, "items": { "type": "string", "maxLength": 80 } }, "tags": { "type": "array", "maxItems": 30, "items": { "type": "string", "maxLength": 80 } }, "properties": { "type": "array", "maxItems": 30, "items": { "$ref": "#/components/schemas/ResourcePropertyInput" } }, "expectedUpdatedAt": { "type": "string", "format": "date-time" } } },
      "ResourceMetadataResponse": { "type": "object", "required": ["data"], "properties": { "data": { "type": "object", "required": ["id", "aliases", "tags", "properties", "updatedAt"], "properties": { "id": { "type": "string", "format": "uuid" }, "aliases": { "type": "array", "items": { "type": "string" } }, "tags": { "type": "array", "items": { "type": "string" } }, "properties": { "type": "object", "additionalProperties": true }, "updatedAt": { "type": "string", "format": "date-time" } } } } },
      "ResourceMention": { "type": "object", "required": ["id", "sourceResourceId", "sourceVersionId", "targetResourceId", "rawTarget", "displayText", "anchor", "startOffset", "endOffset", "resolution"], "properties": { "id": { "type": "string", "format": "uuid" }, "sourceResourceId": { "type": "string", "format": "uuid" }, "sourceVersionId": { "type": "string", "format": "uuid" }, "targetResourceId": { "type": ["string", "null"], "format": "uuid" }, "rawTarget": { "type": "string" }, "displayText": { "type": ["string", "null"] }, "anchor": { "type": ["string", "null"] }, "startOffset": { "type": "integer" }, "endOffset": { "type": "integer" }, "resolution": { "type": "string", "enum": ["linked", "unresolved", "ambiguous"] }, "targetTitle": { "type": ["string", "null"] }, "sourceTitle": { "type": "string" } } },
      "ResourceReferencesResponse": { "type": "object", "required": ["data"], "properties": { "data": { "type": "object", "required": ["outgoing", "incoming"], "properties": { "outgoing": { "type": "array", "items": { "$ref": "#/components/schemas/ResourceMention" } }, "incoming": { "type": "array", "items": { "$ref": "#/components/schemas/ResourceMention" } } } } } },
      "ResourceResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Resource"
          }
        }
      },
      "ResourceListResponse": {
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Resource"
            }
          },
          "meta": {
            "type": "object",
            "required": [
              "nextCursor"
            ],
            "properties": {
              "nextCursor": {
                "type": [
                  "string",
                  "null"
                ]
              }
            }
          }
        }
      },
      "ResourceAccess": {
        "type": "object",
        "required": [
          "url",
          "expiresIn",
          "mode"
        ],
        "properties": {
          "url": {
            "type": "string",
            "format": "uri"
          },
          "expiresIn": {
            "type": "integer"
          },
          "mode": {
            "type": "string",
            "enum": [
              "inline",
              "download"
            ]
          }
        }
      },
      "ResourceAccessResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ResourceAccess"
          }
        }
      },
      "AccessibilityInput": {
        "type": "object",
        "required": [
          "text"
        ],
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 2000
          }
        }
      },
...
    }
  }
}